Implemented logging & beautified code
This logging system is very precise (by logging almost all activity on the server and saving events in a hierarchy where one event is caused by another) but it might even be too much since it's generating a shit ton of data. But I'll keep it for now.

Also, made the code better & more modular by using the express Router.